LAVA-852 - LXC fastboot support for persistent devices in V2

+++ lava_dispatcher/pipeline/protocols/lxc.py
@@ -19,6 +19,7 @@
 # with this program; if not, see <http://www.gnu.org/licenses>.
 
 
+import os
 import pexpect
 import logging
 from lava_dispatcher.pipeline.connection import Protocol
@@ -27,7 +28,11 @@ from lava_dispatcher.pipeline.action import (
     Timeout,
 )
 from lava_dispatcher.pipeline.shell import ShellCommand
-from lava_dispatcher.pipeline.utils.constants import LAVA_LXC_TIMEOUT
+from lava_dispatcher.pipeline.utils.constants import (
+    LAVA_LXC_TIMEOUT,
+    LXC_PATH,
+)
+from lava_dispatcher.pipeline.utils.filesystem import lxc_path
 
 
 class LxcProtocol(Protocol):
@@ -39,8 +44,13 @@ class LxcProtocol(Protocol):
     def __init__(self, parameters, job_id):
         super(LxcProtocol, self).__init__(parameters, job_id)
         self.system_timeout = Timeout('system', LAVA_LXC_TIMEOUT)
-        self.lxc_name = '-'.join([parameters['protocols'][self.name]['name'],
-                                  str(job_id)])
+        self.persistence = parameters['protocols'][self.name].get('persist',
+                                                                  False)
+        if self.persistence:
+            self.lxc_name = parameters['protocols'][self.name]['name']
+        else:
+            self.lxc_name = '-'.join(
+                [parameters['protocols'][self.name]['name'], str(job_id)])
         self.lxc_dist = parameters['protocols'][self.name]['distribution']
         self.lxc_release = parameters['protocols'][self.name]['release']
         self.lxc_arch = parameters['protocols'][self.name]['arch']
@@ -52,6 +62,9 @@ class LxcProtocol(Protocol):
             'security_mirror', None)
         self.verbose = parameters['protocols'][self.name].get('verbose', False)
         self.fastboot_reboot = parameters.get('reboot_to_fastboot', True)
+        self.custom_lxc_path = False
+        if not LXC_PATH == lxc_path(parameters['dispatcher']):
+            self.custom_lxc_path = True
         self.logger = logging.getLogger('dispatcher')
 
     @classmethod
@@ -112,16 +125,34 @@ class LxcProtocol(Protocol):
         # execute the command.
         shell.expect(pexpect.EOF)
         if not shell.exitstatus:
-            self.logger.info("%s protocol: %s exists, proceed to destroy",
-                             self.name, self.lxc_name)
-            cmd = "lxc-destroy -n {0} -f".format(self.lxc_name)
-            self.logger.debug("%s protocol: executing '%s'", self.name, cmd)
+            self.logger.info("%s protocol: %s exists", self.name,
+                             self.lxc_name)
+            # Check if the container should persist
+            if self.persistence:
+                self.logger.debug("%s protocol: issue stop, to persist",
+                                  self.name)
+                cmd = "lxc-stop -n {0} -k".format(self.lxc_name)
+                self.logger.debug("%s protocol: executing '%s'", self.name,
+                                  cmd)
+            else:
+                self.logger.debug("%s protocol: destroy", self.name)
+                if self.custom_lxc_path:
+                    abs_path = os.path.realpath(os.path.join(LXC_PATH,
+                                                             self.lxc_name))
+                    cmd = "lxc-destroy -n {0} -f -P {1}".format(
+                        self.lxc_name, os.path.dirname(abs_path))
+                else:
+                    cmd = "lxc-destroy -n {0} -f".format(self.lxc_name)
+                self.logger.debug("%s protocol: executing '%s'", self.name,
+                                  cmd)
             shell = ShellCommand("%s\n" % cmd, self.system_timeout,
                                  logger=self.logger)
             # execute the command.
             shell.expect(pexpect.EOF)
             if shell.exitstatus:
-                raise InfrastructureError("%s command exited %d: %s" % (cmd,
-                                                                        shell.exitstatus,
-                                                                        shell.readlines()))
+                raise InfrastructureError(
+                    "%s command exited %d: %s" % (cmd, shell.exitstatus,
+                                                  shell.readlines()))
+            if self.custom_lxc_path and not self.persistence:
+                os.remove(os.path.join(LXC_PATH, self.lxc_name))
         self.logger.debug("%s protocol finalised.", self.name)
